pyproject.toml
[tool.poetry]
name = "drf-task"
version = "0.1.0"
description = ""
authors = ["aslan <baristan1aslan@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_task"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.4"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.4"
djoser = "^2.1.0"
djangorestframework-simplejwt = "4.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

